Extract mocking fixtures into a separate source set
