+++ src/enterpriseCompatibilityTest/java/com/gradle/ccud/adapters/enterprise/proxies/GradleEnterpriseBuildCacheProxyTest.java
@@ -5,6 +5,8 @@ import com.gradle.enterprise.gradleplugin.GradleEnterpriseBuildCache;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -12,7 +14,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class GradleEnterpriseBuildCacheProxyTest extends BaseProxyTest {
+@ExtendWith(MockitoExtension.class)
+public class GradleEnterpriseBuildCacheProxyTest {
 
     private GradleEnterpriseBuildCache cache;
     private GradleEnterpriseBuildCacheProxy proxy;
