+++ src/enterpriseCompatibilityTest/java/com/gradle/ccud/adapters/enterprise/proxies/GradleEnterpriseExtensionProxyTest.java
@@ -6,6 +6,8 @@ import com.gradle.scan.plugin.BuildScanExtension;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -13,7 +15,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-class GradleEnterpriseExtensionProxyTest extends BaseProxyTest {
+@ExtendWith(MockitoExtension.class)
+class GradleEnterpriseExtensionProxyTest {
 
     private GradleEnterpriseExtension extension;
     private GradleEnterpriseExtensionProxy proxy;
